zkDSL compiler: split lowering by the question each part answers
lower.rs was 3,124 lines at the start of this work and had no internal boundary.
It is now 1,824, with three more modules beside the eval one split out earlier.
Pure moves: no digest changes and the guest is unchanged.

The split is by invariant, not by size, because each of the four has one worth
stating once instead of rediscovering. eval asks what an expression is worth
before anything runs, and every function in it takes &self and emits nothing.
mem asks which cell a name means and what writing to it costs: every index is
bounds-checked in every position, and a deferred store is a value fact rather
than an instruction. call is the boundary where caller and callee must agree on
the arity, since they place the return area from their own idea of it.
builtins is the precompile and the hints, the two places a value arrives without
an instruction computing it.

Three of those sentences are the bugs found this week, written where the next
person will read them: the unchecked store index, the missing arity check, and
the hint destination that has to be materialised first.
