+++ crates/lean_compiler/src/lower.rs
@@ -1,12 +1,34 @@
 //! Lowering: each function AST is compiled to a sequence of intermediate
 //! [`LOp`] instructions (fp-relative offsets, backpatched jump targets).
+//!
+//! This file holds the walk itself, control flow, instruction emission, and
+//! [`Scope`]. The rest is split by the question it answers, because each of the
+//! four has an invariant worth stating once rather than rediscovering:
+//!
+//! - [`mod@eval`] asks what an expression is worth before anything runs. Every
+//!   function there takes `&self` and emits nothing, which is what lets a caller
+//!   ask without paying for the answer.
+//! - [`mod@mem`] asks which cell a name means, and what writing to it costs.
+//!   Every index is bounds-checked, in every position, and a deferred store is a
+//!   value fact rather than an instruction.
+//! - [`mod@call`] is the call boundary. Caller and callee must agree on the
+//!   arity, because they place the return area from their own idea of it.
+//! - [`mod@builtins`] is the precompile and the hints: the two places a value
+//!   arrives without an instruction computing it.
+//!
+//! [`Scope`] is what a name means HERE, and it reverts at a branch join, so a
+//! cell whose `SET` sits inside a branch is never trusted outside it.
 
 use super::*;
 use crate::filler::FillerOp;
 use lean_vm::cpu::filler::Block;
 use std::collections::HashSet;
 
+mod builtins;
+mod call;
 mod eval;
+mod mem;
+use call::ret_binding;
 use eval::field_pow;
 /// [`FnLower::specialized_body`]'s pieces: runtime param names, runtime args,
 /// the `Const`-substituted body, and the callee's return arity.
@@ -98,17 +120,6 @@ enum Binding {
     Stack(Off, u32),
     Gaddr(GAddr),
     FConst(F192),
-}
-
-/// How an inlined tail return binds in the caller: a `StackBuf` run and a folded
-/// g-address alias at zero copies, while anything else (a plain scalar, or a real
-/// call, which records no [`RetBind`]) takes the destination cell it wrote.
-fn ret_binding(b: Option<RetBind>, dst: Off) -> Binding {
-    match b {
-        Some(RetBind::Stack(base, size)) => Binding::Stack(base, size),
-        Some(RetBind::Gaddr(ga)) => Binding::Gaddr(ga),
-        _ => Binding::Scalar(dst),
-    }
 }
 
 /// Everything a runtime branch may not have executed: the name bindings, plus
@@ -311,25 +322,6 @@ impl FnLower<'_> {
         });
     }
 
-    /// Prepare a stack run that a consumer is about to name by its *physical*
-    /// cells: a `BLAKE2s` output, or a hint destination. Those consumers do not go
-    /// through [`Self::word_src`], so a cell still carrying a deferred alias would
-    /// have the consumer's write land where nothing reads it, and the equality
-    /// assertion the source wrote would be gone. Materializing the alias first puts
-    /// a real value in the cell, which is what turns the consumer's write back into
-    /// that assertion; marking the run `phys` covers the other order, where the
-    /// store comes after the consumer.
-    fn materialize_run(&mut self, base: Off, len: u32) {
-        for o in base..base + len {
-            if self.alias.contains_key(&o) {
-                let src = self.word_src(o);
-                self.alias.remove(&o);
-                self.copy(src, o);
-            }
-            self.phys.insert(o);
-        }
-    }
-
     fn set(&mut self, o: Off, k: KVal) {
         self.emit(LOp::Set { o, k });
     }
@@ -430,102 +422,6 @@ impl FnLower<'_> {
     /// in this scope XORs into.
     fn zero(&mut self) -> Off {
         self.const_cell(F192::ZERO)
-    }
-
-    fn default_blake2s_cv(&mut self) -> Off {
-        if let Some(o) = self.scope.blake2s_iv {
-            return o;
-        }
-        let o = self.alloc_stack(2);
-        for (k, value) in lean_vm::hash_flock::IV_CELLS.into_iter().enumerate() {
-            self.set_const(o + k as u32, value);
-            self.scope
-                .const_cells
-                .entry([value.c0, value.c1, value.c2])
-                .or_insert(o + k as u32);
-        }
-        self.scope.blake2s_iv = Some(o);
-        o
-    }
-
-    /// A stack store `sa[k] = val` whose value is a plain copy or a zero, which we
-    /// defer as an [`Alias`] (forwarded at use) instead of emitting.
-    fn copy_alias(&self, val: &Expr) -> Option<Alias> {
-        match val {
-            // A live var / stack cell aliases to that cell directly (no new
-            // material); anything else that is a compile-time constant defers
-            // to the pooled const cell.
-            Expr::Var(v) if self.scope.var(v).is_some() => self.scope.var(v).map(Alias::Cell),
-            Expr::Index(arr, idx) => {
-                // Bounds-checked here too. This is the ONE stack index that was
-                // not, and it is on the hot path: `sa[i] = sb[j]` and every
-                // element of a list literal reach it. Unchecked, `c[0] = a[2]`
-                // on a `StackBuf(2)` aliased the next buffer's first cell and
-                // its assert passed, while a further index named a cell nothing
-                // writes, so a prover chose it and the assert was vacuous. The
-                // identical read in expression position was rejected, so this
-                // was two spellings with two meanings.
-                let (base, size) = self.stack_of(arr)?;
-                let k = self.try_const_index(idx)?;
-                if k >= size {
-                    self.fail(format!("stack index {k} out of bounds (size {size})"))
-                }
-                Some(Alias::Cell(base + k))
-            }
-            _ => self.try_field_const(val).map(Alias::Const),
-        }
-    }
-
-    /// Write `val` into the stack cell `dst`. A plain copy or a constant is
-    /// deferred as an [`Alias`] and forwarded at its uses (write-once, so the
-    /// source cell keeps its value): the assembling `MUL`/`SET` is never emitted.
-    fn stack_store(&mut self, dst: Off, val: &Expr) {
-        // Deferring is only sound while nothing else has given `dst` a value. Once
-        // something has, the store IS the write-once equality assertion of
-        // `zkDSL.md` §Memory, so it has to be emitted: an alias would silently
-        // redirect every later read to the source and drop the assertion. This is
-        // what makes `s[k] = <checked value>` pin a hint, and what makes a
-        // pre-written `blake2s` output assert the digest.
-        let aliased = self.alias.contains_key(&dst);
-        if !aliased
-            && !self.phys.contains(&dst)
-            && let Some(a) = self.copy_alias(val)
-        {
-            // Record the end of the `Cell` chain, not its head. A `Cell` alias is
-            // otherwise free to point at another alias, and two of them can close
-            // a loop (`s[0] = s[1]` then `s[1] = s[0]`, or the one-line
-            // `s[0] = s[0]`) that `word_src` then walks forever. What rules a
-            // cycle out is that the recorded source has no outgoing `Cell` edge AT
-            // THE MOMENT IT IS RECORDED; it may acquire one later, so chains still
-            // grow and `word_src`'s bound is load-bearing, not decoration.
-            //
-            // [`Self::cell_src`] rather than [`Self::word_src`] on purpose: the
-            // latter resolves a trailing `Const` through `zero`/`const_cell`,
-            // which EMITS a `SET`. Paying it here rather than at the use would
-            // make `sa[k] = other` cost an instruction, which `zkDSL.md`
-            // §Variables promises it does not.
-            let a = match a {
-                Alias::Cell(src) => Alias::Cell(self.cell_src(src)),
-                other => other,
-            };
-            // A cell aliased to itself is the whole of `sb[k] = sb[k]`: write-once
-            // makes a second write of the same value a no-op, so emit nothing.
-            if a == Alias::Cell(dst) {
-                return;
-            }
-            self.alias.insert(dst, a);
-            return;
-        }
-        if aliased {
-            // Give the cell the value it already stood for, so the store below is a
-            // second write of that cell and therefore the assertion. Without this the
-            // second alias would simply replace the first and the two values would
-            // never meet.
-            let src = self.word_src(dst);
-            self.alias.remove(&dst);
-            self.copy(src, dst);
-        }
-        self.expr_into(val, dst);
     }
 
     /// Terminate `main`: jump to the halt sentinel `g^{B-1}` with `fp = g^0`.
@@ -802,89 +698,6 @@ impl FnLower<'_> {
         self.bind_join(names, &rcells);
     }
 
-    /// `names = match_range(log(x), …, lambda k: f(args, k))` fused: the arms all
-    /// call one of `callees` (specializations sharing the arg/return layout) with
-    /// the same runtime `args`, so build the callee frame **once** and let the
-    /// dispatch jump straight into the selected entry, which returns to the join.
-    /// Each taken arm is then just the trampoline's `SET entry; JUMP`: no
-    /// per-arm frame setup, call, or return jump.
-    fn lower_dispatched_call(&mut self, names: &[String], x: &Expr, callees: &[String], rt_args: &[Expr]) {
-        let n_args = rt_args.len() as u32;
-        // The join below reads one return cell per bound name, so every callee has
-        // to declare exactly that many. Unchecked, a name past a callee's arity
-        // `DEREF`s a frame offset nothing on that path writes, and since the shared
-        // frame is sized to the LARGEST callee the offset exists: the surplus name
-        // binds a prover-chosen word. The non-fused path enforces this
-        // ([`Self::call_into`]), so leaving it out here means one source is rejected
-        // by one lowering of `match_range` and silently miscompiled by the other.
-        for callee in callees {
-            // Arguments for the same reason as returns below: the shared frame
-            // is sized to the largest callee, so a callee expecting more than
-            // the arms supply reads a cell that exists and nothing writes.
-            if let Some(want) = self.arity_of(callee)
-                && want != rt_args.len()
-            {
-                let plural = if want == 1 { "argument" } else { "arguments" };
-                self.fail(format!(
-                    "`{callee}` takes {want} {plural}, dispatched call passes {}",
-                    rt_args.len()
-                ))
-            }
-            let Some(shapes) = self.return_shapes_of(callee) else {
-                continue;
-            };
-            if shapes.len() != names.len() {
-                self.fail(format!(
-                    "`{callee}` returns {} values, dispatched call binds {}",
-                    shapes.len(),
-                    names.len()
-                ))
-            };
-            if shapes.iter().any(|s| *s != ReturnShape::Scalar) {
-                self.fail(format!(
-                    "`{callee}`: a multi-cell StackBuf return cannot cross a dispatched join"
-                ))
-            };
-        }
-        let rcells: Vec<Off> = names.iter().map(|_| self.fresh()).collect();
-
-        // Shared callee frame: args, retfp, and retpc = the join (so the callee
-        // returns straight past the dispatch). Evaluated once.
-        let arg_offs: Vec<Off> = rt_args.iter().map(|a| self.expr(a)).collect();
-        let xo = self.expr(x);
-        let one = self.one();
-        let sfp = self.self_fp();
-
-        let nfp = self.fresh();
-        self.pending.push(Hint::AllocFrameMax {
-            ptr: nfp,
-            callees: callees.to_vec(),
-        });
-        for (i, &ao) in arg_offs.iter().enumerate() {
-            self.deref(nfp, Abi::arg(i as u32), ao, DerefMode::Cell);
-        }
-        self.deref(nfp, Abi::RET_FP, 0, DerefMode::Fp);
-        let join_cell = self.fresh();
-        let join_set = self.code.len();
-        self.set(join_cell, KVal::Local(0)); // patched: the join pc
-        self.deref(nfp, Abi::RET_PC, join_cell, DerefMode::Cell); // retpc = join
-
-        let kset = self.emit_dispatch(xo, one, sfp);
-
-        // Trampoline: slot j enters `callees[j]` with fp = nfp; the callee's own
-        // `return` jumps to retpc (the join) in the caller frame.
-        self.patch_local(kset, self.code.len());
-        self.emit_slots(callees.len(), one, nfp, |j| KVal::Entry(callees[j].clone()));
-
-        // Join: read the return values (written by whichever callee ran).
-        self.patch_local(join_set, self.code.len());
-        for (i, &r) in rcells.iter().enumerate() {
-            self.deref(nfp, Abi::ret(n_args, i as u32), r, DerefMode::Cell);
-        }
-
-        self.bind_join(names, &rcells);
-    }
-
     /// The two-jump dispatch itself: `d = g^T · x²` names slot `x` of the
     /// two-instruction trampoline table at bytecode base `T`. Returns the index
     /// of the `SET` holding `T`, for the caller to patch once the table's
@@ -1082,90 +895,6 @@ impl FnLower<'_> {
     /// would now land on all of them.
     fn bound_cell(&mut self, k: u64) -> Off {
         self.const_cell(g_pow_u128((k - 1) as u128).into())
-    }
-
-    /// Resolve an expression naming a run of consecutive cells: a whole
-    /// `StackBuf`, a `StackBuf` slice, a `HeapBuf` slice with compile-time
-    /// bounds, or a runtime-start heap slice `buf[i:i + k]` (whose length is
-    /// the only thing its bounds reveal). Heap runs fold the buffer's symbolic
-    /// shift and the slice start into the pointer offset.
-    fn cell_run(&mut self, e: &Expr) -> CellRun {
-        match e {
-            Expr::Var(_) => {
-                let (base, len) = self
-                    .stack_of(e)
-                    .expect("only a StackBuf names a cell run unsliced; slice a HeapBuf: `buf[lo:lo + k]`");
-                CellRun::Stack { base, len }
-            }
-            Expr::Slice(arr, lo, hi) => match (self.try_const_index(lo), self.try_const_index(hi)) {
-                // Compile-time bounds: integer cell indexes `lo..hi` (frame
-                // offsets for a stack, g-power exponents for the heap).
-                (Some(lo), Some(hi)) => {
-                    if lo >= hi {
-                        self.fail(format!("empty slice {lo}:{hi}"))
-                    };
-                    let len = hi - lo;
-                    if let Some((base, size)) = self.stack_of(arr) {
-                        if hi > size {
-                            self.fail(format!("slice {lo}:{hi} out of bounds (StackBuf size {size})"))
-                        };
-                        CellRun::Stack { base: base + lo, len }
-                    } else {
-                        self.check_heap_bound(arr, lo as u128, len as u128);
-                        let (ptr, lo) = self.heap_base(arr, lo as u128);
-                        CellRun::Heap { ptr, lo, len }
-                    }
-                }
-                // Runtime start (heap only): `buf[i:i + k]` with a runtime
-                // g-power index `i` names the cells `buf·i·g^j`, j < k. The
-                // `hi` bound cannot be evaluated, only shape-checked against
-                // `lo`. One MUL folds `i` into the pointer.
-                _ => {
-                    if self.stack_of(arr).is_some() {
-                        self.fail(
-                            "a StackBuf slice needs compile-time bounds (frame offsets are baked into the bytecode)",
-                        )
-                    };
-                    let k = plus_k(lo, hi).unwrap_or_else(|| {
-                        self.fail(format!("a runtime slice must be `buf[i:i + k]`, got `{lo:?}:{hi:?}`"))
-                    });
-                    let len = u32::try_from(k).expect("slice length overflows u32");
-                    if len == 0 {
-                        self.fail(format!("a runtime slice `{lo:?}:{hi:?}` has length 0, so it names no cell"))
-                    };
-                    // `heap_addr` bounds-checks ONE cell. A start that folds
-                    // (`GEN ** k`, or a name bound to one) arrives here because
-                    // it is not an INTEGER, yet its offset is known, so the run's
-                    // length has to be checked here or it never is: the same run
-                    // written with integer bounds was rejected, while this
-                    // spelling let a `hint_witness` write past the buffer into
-                    // the next one.
-                    if let Some(GAddr { base: None, exp, .. }) = self.gaddr_of(lo) {
-                        self.check_heap_bound(arr, exp, u128::from(len));
-                    }
-                    let (ptr, lo) = self.heap_addr(arr, lo);
-                    CellRun::Heap { ptr, lo, len }
-                }
-            },
-            other => self.fail(format!(
-                "expected a StackBuf, a StackBuf slice, or a HeapBuf slice, got `{other:?}`"
-            )),
-        }
-    }
-
-    /// `hint_witness(dest, "name")`: resolve `dest` to a run of cells and
-    /// queue the witness-fill hint (no instructions: the values are written
-    /// by the runner before the next instruction executes, unconstrained).
-    fn lower_hint_witness(&mut self, dest: &Expr, name: &str) {
-        let name = name.to_string();
-        let hint = match self.cell_run(dest) {
-            CellRun::Stack { base, len } => {
-                self.materialize_run(base, len);
-                RHint::WitnessStack { name, base, len }
-            }
-            CellRun::Heap { ptr, lo, len } => RHint::WitnessHeap { name, ptr, lo, len },
-        };
-        self.pending.push(Hint::Resolved(hint));
     }
 
     /// `assert log x < log GEN ** k`: the 3-cycle range check *in the
@@ -1384,87 +1113,6 @@ impl FnLower<'_> {
         }
     }
 
-    /// Allocate `n` *consecutive* fresh frame cells (a stack run), returning the
-    /// base. Nothing else may `fresh()` between them, so they stay adjacent.
-    fn alloc_stack(&mut self, n: u32) -> Off {
-        let base = self.next;
-        self.next += n;
-        // A run allocated after the escape is just as reachable as one before it,
-        // so it is sealed on the spot and never needs recording.
-        if let Some(runs) = &mut self.unsealed_runs {
-            runs.push((base, n));
-        } else {
-            self.seal_run(base, n);
-        }
-        base
-    }
-
-    /// Give up on seeing writes to a frame run: materialize any deferred alias on
-    /// it, mark its cells `phys` so a later store stays the write-once equality
-    /// assertion rather than an alias, and hand it to CSE as untouchable.
-    fn seal_run(&mut self, base: Off, len: u32) {
-        self.materialize_run(base, len);
-        self.opaque_runs.push((base, len));
-    }
-
-    /// A computed-advice bit buffer's destination ([`BitsDest`]). Not
-    /// [`Self::cell_run`]: these builtins take a bare `HeapBuf` and carry the
-    /// length in `nbits`, where a cell run would demand a slice.
-    fn bits_dest(&mut self, e: &Expr, nbits: u32, what: &str) -> BitsDest {
-        match self.stack_of(e) {
-            Some((base, len)) => {
-                if len < nbits {
-                    self.fail(format!(
-                        "{what} needs {nbits} cells, its StackBuf destination has {len}"
-                    ))
-                };
-                // The hint names the physical cells, so a deferred alias sitting on
-                // one would win every later read (as for `hint_f192_limbs`).
-                self.materialize_run(base, nbits);
-                BitsDest::Stack(base)
-            }
-            None => BitsDest::Heap(self.expr(e)),
-        }
-    }
-
-    /// If `e` names a `StackBuf` variable, its `(base, size)`.
-    fn stack_of(&self, e: &Expr) -> Option<(Off, u32)> {
-        match e {
-            Expr::Var(v) => self.scope.stack(v),
-            _ => None,
-        }
-    }
-
-    /// `addr(sb)`: the g-address `fp·g^base` of a `StackBuf`'s first cell, so a
-    /// frame run can be pointed at (indexed at runtime, or handed to a callee)
-    /// while its own accesses stay direct frame cells. Materializing `fp` is the
-    /// ISA's one cost here, amortized per function ([`Self::self_fp`]); the
-    /// address is a folded [`GAddr`], so `addr(sb) * GEN ** k` stays virtual.
-    fn stack_addr(&mut self, args: &[Expr]) -> GAddr {
-        if args.len() != 1 {
-            self.fail("addr(buf) takes one StackBuf")
-        };
-        let (base, len) = self.stack_of(&args[0]).expect("addr() takes a StackBuf");
-        // Once an address escapes, a write through it is a `DEREF` that names no
-        // frame cell the lowerer can see, so every run has to be treated as
-        // already valued: `phys` keeps a later `sb[k] = v` the write-once
-        // equality assertion `zkDSL.md` §Memory promises instead of a deferred
-        // alias that would drop it, and `opaque_runs` keeps CSE off its cells (a
-        // store folded into a canonical elsewhere would leave the cell unwritten,
-        // hence prover-chosen). Every run, not just this one: the pointer is a
-        // frame address, and one off-by-one in a callee reaches the next run.
-        if let Some(runs) = self.unsealed_runs.take() {
-            for (b, l) in runs {
-                self.seal_run(b, l);
-            }
-        }
-        GAddr {
-            base: Some(self.self_fp()),
-            exp: base as u128,
-            run: Some((base, len)),
-        }
-    }
-
     /// `base ** e` (non-`GEN` base, compile-time exponent `e`): a fully-constant
     /// base folds to one `SET`; a runtime base is raised by square-and-multiply.
     fn pow_expr(&mut self, b: &Expr, e: &Expr) -> Off {
@@ -1497,39 +1145,6 @@ impl FnLower<'_> {
         acc
     }
 
-    /// Resolve a `blake2s` operand: a [`Self::cell_run`] pinned to exactly 2
-    /// cells, a 256-bit value being two 128-bit cells. Stack operands are used
-    /// in place; heap operands must be bridged through the stack, since
-    /// `BLAKE2s` addresses only frame cells (see [`Self::blake2s_input`]).
-    fn blake2s_operand(&mut self, e: &Expr) -> CellRun {
-        let run = self.cell_run(e);
-        if run.cells() != 2 {
-            self.fail("a blake2s operand must span exactly 2 cells (two 128-bit words); slice a larger buffer: `buf[lo:lo + 2]`")
-        };
-        run
-    }
-
-    /// A `blake2s` *input* operand as its two independently-addressed 128-bit
-    /// chunk bases (each chunk is ONE 128-bit cell): stack runs in place; a heap
-    /// slice is pulled into a fresh stack pair first, one `DEREF` per cell
-    /// (`m[ptr·g^{lo+k}] == m[fp+t+k]`, the `β` immediate doing the pointer
-    /// offset). The heap cells must already be written.
-    fn blake2s_input(&mut self, e: &Expr) -> [Off; 2] {
-        match self.blake2s_operand(e) {
-            // A stack operand: the two chunk cells are `o, o+1`; forward each
-            // cell's real source where known (a copy or a zero), so a hash of
-            // non-adjacent values needs no assembling copies.
-            CellRun::Stack { base, .. } => [self.word_src(base), self.word_src(base + 1)],
-            CellRun::Heap { ptr, lo, .. } => {
-                let t = self.alloc_stack(2);
-                for k in 0..2 {
-                    self.deref(ptr, lo + k, t + k, DerefMode::Cell);
-                }
-                [t, t + 1]
-            }
-        }
-    }
-
     /// A BLAKE2s chaining value must occupy two consecutive frame cells because
     /// the opcode carries one base offset for both words. Preserve a genuine
     /// consecutive pair, including a heap pair already bridged by
@@ -1544,39 +1159,6 @@ impl FnLower<'_> {
         self.copy(pair[0], cv);
         self.copy(pair[1], cv + 1);
         cv
-    }
-
-    /// The cell holding the value of stack cell `o`, following a recorded copy /
-    /// zero alias to its real source. Returns `o` when it holds a genuine value.
-    /// Follow `Cell` alias hops to the cell that ends the chain, emitting nothing.
-    /// This is what [`Self::stack_store`] records, so a recorded source never has
-    /// an outgoing `Cell` edge and no cycle can form.
-    fn cell_src(&self, o: Off) -> Off {
-        let mut cur = o;
-        for _ in 0..=self.next {
-            match self.alias.get(&cur) {
-                Some(Alias::Cell(s)) => cur = *s,
-                _ => return cur,
-            }
-        }
-        self.fail(format!("alias cycle at frame cell {o}"));
-    }
-
-    fn word_src(&mut self, o: Off) -> Off {
-        // Bounded because a chain can still GROW: a recorded source is unaliased
-        // when it is recorded, but nothing stops it acquiring an alias afterwards.
-        // A cycle here is a compiler that never returns and prints nothing, so
-        // more hops than there are frame cells is a hard error, not a hang.
-        let mut cur = o;
-        for _ in 0..=self.next {
-            match self.alias.get(&cur).copied() {
-                Some(Alias::Cell(s)) => cur = s,
-                Some(Alias::Const(v)) if v.is_zero() => return self.zero(),
-                Some(Alias::Const(v)) => return self.const_cell(v),
-                None => return cur,
-            }
-        }
-        self.fail(format!("alias cycle at frame cell {o}"));
     }
 
     /// Evaluate `e` writing its value straight into cell `dst`, with no temporary +
@@ -1633,281 +1215,6 @@ impl FnLower<'_> {
         }
     }
 
-    /// Resolve a heap access `arr[idx]` to a `DEREF`-ready pair: a cell
-    /// holding a pointer `p` and a compile-time exponent `beta`, the accessed
-    /// cell being `m[p·g^beta]` (heap addressing in the exponent: cell `g^k`
-    /// of the buffer sits at `arr·g^k`). The fallback of [`Self::heap_addr`],
-    /// which has already folded away a wholly constant index: here a constant
-    /// g-power *factor* still goes into the `beta` immediate, so only the
-    /// runtime factor costs a pointer `MUL`.
-    fn array_ptr(&mut self, arr: &Expr, idx: &Expr) -> (Off, u32) {
-        // `buf[r * GEN ** k]` (either factor order): beta takes the constant,
-        // the pointer MUL takes only the runtime factor `r`.
-        if let Expr::Mul(a, b) = idx {
-            for (c, r) in [(a, b), (b, a)] {
-                if let Some(k) = self.const_gpow(c) {
-                    let (la, lr) = (self.expr(arr), self.expr(r));
-                    let ptr = self.fresh();
-                    self.emit(LOp::Mul { a: la, b: lr, c: ptr });
-                    return (ptr, k);
-                }
-            }
-        }
-        let (la, li) = (self.expr(arr), self.expr(idx));
-        let ptr = self.fresh();
-        self.emit(LOp::Mul { a: la, b: li, c: ptr });
-        (ptr, 0)
-    }
-
-    /// Realize a [`GAddr`] into a frame cell holding its value: a constant is one
-    /// `SET`; a base with no shift is already that cell; a shifted base is a
-    /// `SET`+`MUL`.
-    fn materialize(&mut self, ga: GAddr) -> Off {
-        match ga {
-            GAddr {
-                base: Some(c), exp: 0, ..
-            } => c,
-            GAddr { base, exp, .. } => {
-                let k = self.fresh();
-                self.set_const(k, g_pow_u128(exp).into());
-                let Some(c) = base else { return k };
-                let o = self.fresh();
-                self.emit(LOp::Mul { a: c, b: k, c: o });
-                o
-            }
-        }
-    }
-
-    /// Compile-time bounds check: when `arr` resolves to a sized `HeapBuf`
-    /// (directly or through shifted aliases) and the whole index is the
-    /// compile-time exponent `exp`, reject `exp + span > size`. Runtime
-    /// indices are not checked (their value is not known here).
-    fn check_heap_bound(&self, arr: &Expr, extra: u128, span: u128) {
-        let Some(ga) = self.gaddr_of(arr) else { return };
-        let (Some(base), Some(exp)) = (ga.base, ga.exp.checked_add(extra)) else {
-            return;
-        };
-        // A frame pointer from `addr(sb)`: `exp` is an absolute frame offset and
-        // `base` is the shared `fp` cell, so the run comes from the address's own
-        // provenance rather than from `heap_sizes`.
-        if let Some((start, len)) = ga.run {
-            let (start, len) = (start as u128, len as u128);
-            if exp < start || exp + span > start + len {
-                let off = exp.saturating_sub(start); // `exp < start` is unreachable: gmul only adds
-                let what = if span == 1 {
-                    format!("index {off}")
-                } else {
-                    format!("slice {off}:{}", off + span)
-                };
-                self.fail(format!(
-                    "frame {what} out of bounds for the StackBuf({len}) named by `addr`"
-                ));
-            }
-            return;
-        }
-        let Some(&size) = self.heap_sizes.get(&base) else {
-            return;
-        };
-        if exp + span > size {
-            let name = self
-                .scope
-                .names
-                .iter()
-                .find(|(_, b)| matches!(b.val, Binding::Scalar(c) if c == base))
-                .map(|(n, _)| n.as_str())
-                .unwrap_or("?");
-            if span == 1 {
-                self.fail(format!(
-                    "heap index {exp} out of bounds for `{name}` (HeapBuf size {size})"
-                ));
-            }
-            self.fail(format!(
-                "heap slice {exp}:{} out of bounds for `{name}` (HeapBuf size {size})",
-                exp + span
-            ));
-        }
-    }
-
-    /// Address `arr·g^extra` as `(base_cell, β)`, folding `arr`'s symbolic shift
-    /// and the constant `extra` into `β`. Falls back to a materialized pointer
-    /// (`β = 0`) when there is no runtime base or the offset exceeds [`FOLD_MAX`].
-    fn heap_base(&mut self, arr: &Expr, extra: u128) -> (Off, u32) {
-        self.check_heap_bound(arr, extra, 1);
-        if let Some(ga) = self.gaddr_of(arr)
-            && let (Some(base), Some(exp)) = (ga.base, ga.exp.checked_add(extra))
-            && exp <= FOLD_MAX
-        {
-            return (base, exp as u32);
-        }
-        let a = self.expr(arr);
-        if extra == 0 {
-            return (a, 0);
-        }
-        let k = self.fresh();
-        self.set_const(k, g_pow_u128(extra).into());
-        let ptr = self.fresh();
-        self.emit(LOp::Mul { a, b: k, c: ptr });
-        (ptr, 0)
-    }
-
-    /// Address `arr[idx]` as `(base_cell, β)`. A constant g-power `idx` folds
-    /// into `β` ([`Self::heap_base`]); a runtime index materializes the pointer.
-    fn heap_addr(&mut self, arr: &Expr, idx: &Expr) -> (Off, u32) {
-        // A compile-time index that is a plain field constant but NOT a
-        // g-power (`buf[0]`, `buf[2]`, an integer unroll var) can never name
-        // a heap cell (cell k lives at `buf · g^k`) and would deref a wild
-        // address at proving time. Reject it here, where the source is known.
-        // A BARE literal index is rejected even now that `gaddr_of` reads `2^k`
-        // as `g^k` in a pointer: `hb[2]` would silently mean cell 1, while slice
-        // bounds stayed integer (`hb[2:4]` starts at cell 2), so one spelling
-        // would name two different cells. An index built from `GEN` is fine, and
-        // `1` is `g^0` either way.
-        let bare_int = matches!(self.try_lit(idx), Some(n) if n != 1);
-        if (bare_int || self.gaddr_of(idx).is_none())
-            && let Some(c) = self.try_field_const(idx)
-        {
-            // Two reasons reach here and they read differently. A bare integer
-            // index may well BE a g-power (4 is g²), and is rejected for being
-            // ambiguous against slice syntax rather than for naming nothing.
-            let why = match self.const_gpow(idx) {
-                Some(k) => format!(
-                    "is a plain integer naming cell {k}, while the slice `buf[n:n + 1]` reads the \
-                     same number as cell n. Write `buf[GEN ** {k}]` and say which you mean"
-                ),
-                None => format!(
-                    "folds to the field constant {:#x}:{:#x}, which is not a g-power, so it names \
-                     no heap cell (did an integer index leak in from a StackBuf conversion?)",
-                    c.c1, c.c0
-                ),
-            };
-            self.fail(format!("heap index {why}"));
-        }
-        match self.gaddr_of(idx) {
-            Some(GAddr { base: None, exp, .. }) => return self.heap_base(arr, exp),
-            // A runtime-base index carrying a constant g-power shift
-            // (`buf[cursor * GEN ** k]`): fold the whole constant part (the
-            // index's shift plus `arr`'s own symbolic shift) into `β`, and
-            // emit ONE pointer multiply instead of materializing g^k.
-            Some(GAddr {
-                base: Some(ib), exp, ..
-            }) => {
-                if let Some(ga) = self.gaddr_of(arr)
-                    && let (Some(ab), Some(total)) = (ga.base, ga.exp.checked_add(exp))
-                    && total <= FOLD_MAX
-                {
-                    let ptr = self.fresh();
-                    self.emit(LOp::Mul { a: ab, b: ib, c: ptr });
-                    return (ptr, total as u32);
-                }
-            }
-            None => {}
-        }
-        // Fall back to the constant-g-power-factor fold (a runtime index still
-        // materializes the pointer `MUL`, with any constant factor in `β`).
-        self.array_ptr(arr, idx)
-    }
-
-    /// Consume the [`RetBind`] a single-value inlined tail return recorded,
-    /// for a call in EXPRESSION position (embedded in arithmetic, a store
-    /// RHS, a single-target match arm): there is no name to alias-bind, so an
-    /// aliased return materializes into a plain cell (free for a var / an
-    /// exp-0 g-address; one `MUL` for a shifted pointer). `dst` is the call's
-    /// destination cell, already written by a real call or a plain-scalar
-    /// return, so it is the fallback.
-    fn take_inline_ret_cell(&mut self, dst: Off) -> Off {
-        match self.inline_stack_ret.take().and_then(|b| b.into_iter().next()) {
-            Some(RetBind::Gaddr(ga)) => self.materialize(ga),
-            Some(RetBind::Stack(base, size)) => {
-                if size != 1 {
-                    self.fail("a multi-cell StackBuf return needs a `let` binding, not an expression use")
-                };
-                // Through `word_src`, like every other read of a stack cell: the body may
-                // have filled this cell with a deferred copy or constant, which emits no
-                // instruction, and the raw cell would then be one no instruction writes.
-                // The `let` consumer follows the alias by taking a `Binding::Stack`
-                // ([`ret_binding`]), and an expression use has to agree with it.
-                self.word_src(base)
-            }
-            _ => dst,
-        }
-    }
-
-    /// Lower a call; returns one caller offset per source-level return value.
-    /// A real-call StackBuf return is flattened into consecutive ABI cells and
-    /// copied into a fresh consecutive run in the caller. `inline_stack_ret`
-    /// describes those logical bindings to the surrounding let/tuple lowering.
-    fn call(&mut self, callee: &str, args: &[Expr], n_ret: usize) -> Vec<Off> {
-        if callee == "blake2s" {
-            self.fail("blake2s is a statement: `blake2s(a, b, out)` writes the digest into the 2-cell stack run `out`")
-        };
-        self.inline_stack_ret = None;
-        if self.defs.get(callee).is_some_and(|d| d.inline) {
-            let dsts: Vec<Off> = (0..n_ret).map(|_| self.fresh()).collect();
-            self.call_into(callee, args, &dsts);
-            return dsts;
-        }
-
-        let shapes = self
-            .defs
-            .get(callee)
-            .map(|d| d.return_shapes.clone())
-            .unwrap_or_else(|| vec![ReturnShape::Scalar; n_ret]);
-        if shapes.len() != n_ret {
-            self.fail(format!(
-                "`{callee}` returns {} values, call binds {n_ret}",
-                shapes.len()
-            ))
-        };
-        let mut logical = Vec::with_capacity(n_ret);
-        let mut physical = Vec::new();
-        let mut binds = Vec::with_capacity(n_ret);
-        for shape in shapes {
-            match shape {
-                ReturnShape::Scalar => {
-                    let dst = self.fresh();
-                    logical.push(dst);
-                    physical.push(dst);
-                    binds.push(RetBind::Scalar);
-                }
-                ReturnShape::StackBuf(size) => {
-                    if size == 0 {
-                        self.fail("a returned StackBuf must not be empty")
-                    };
-                    let base = self.alloc_stack(size);
-                    logical.push(base);
-                    physical.extend(base..base + size);
-                    binds.push(RetBind::Stack(base, size));
-                }
-            }
-        }
-        self.lower_call(callee, args, physical.len(), None, Some(&physical), false);
-        self.inline_stack_ret = Some(binds);
-        logical
-    }
-
-    /// Evaluate `callee(args)` into `dsts`, inlining the callee when it is
-    /// `@inline` ([`Self::try_inline`]), else a real call.
-    fn call_into(&mut self, callee: &str, args: &[Expr], dsts: &[Off]) {
-        if callee == "blake2s" {
-            self.fail("blake2s is a statement, not a value-returning call")
-        };
-        if !self.try_inline(callee, args, dsts) {
-            if let Some(def) = self.defs.get(callee) {
-                if def.return_shapes.len() != dsts.len() {
-                    self.fail(format!(
-                        "`{callee}` returns {} values, call binds {}",
-                        def.return_shapes.len(),
-                        dsts.len()
-                    ))
-                };
-                if def.return_shapes.iter().any(|s| *s != ReturnShape::Scalar) {
-                    self.fail("a normal function's multi-cell StackBuf return needs a `let` binding")
-                };
-            }
-            self.lower_call(callee, args, dsts.len(), None, Some(dsts), false);
-        }
-    }
-
     /// The value a `Const` parameter takes, as the literal that substitutes for
     /// it: a `GEN ** k` argument stays a g-power, everything else must fold to a
     /// compile-time integer (a bound name, a const-array element `DEPTH[lvl]`,
@@ -1920,257 +1227,6 @@ impl FnLower<'_> {
             Expr::GPow(k) => Expr::GPow(*k),
             other => Expr::Lit(self.try_const_index(other)? as u128),
         })
-    }
-
-    /// The runtime params, runtime args, and `Const`-substituted body of a call
-    /// to a user function: the ingredients for inlining. `None` for a builtin or
-    /// unknown callee, an arity mismatch, or an unresolved `Const` argument.
-    fn specialized_body(&self, callee: &str, args: &[Expr]) -> Option<SpecializedBody> {
-        let def = self.defs.get(callee)?;
-        if args.len() != def.params.len() {
-            return None;
-        }
-        let mut body = def.body.clone();
-        let (mut rt_params, mut rt_args) = (Vec::new(), Vec::new());
-        for ((p, &is_const), a) in def.params.iter().zip(&def.const_params).zip(args) {
-            if !is_const {
-                rt_params.push(p.clone());
-                rt_args.push(a.clone());
-                continue;
-            }
-            let c = self.const_arg(a)?;
-            body = subst_stmts(&body, p, &c);
-        }
-        Some((rt_params, rt_args, body, def.n_ret))
-    }
-
-    /// Inline an `@inline` `callee(args)` into the current frame, binding its
-    /// return values straight into `dsts`: no frame setup, no argument/return
-    /// plumbing, no call/return jumps. Returns `false` for a non-`@inline`
-    /// callee (the caller emits a real call). Panics if an `@inline` function
-    /// isn't inlinable ([`body_inlinable`]) or its `Const` args don't resolve.
-    fn try_inline(&mut self, callee: &str, args: &[Expr], dsts: &[Off]) -> bool {
-        if !self.defs.get(callee).is_some_and(|d| d.inline) {
-            return false;
-        }
-        let (params, rt_args, body, n_ret) = self
-            .specialized_body(callee, args)
-            .unwrap_or_else(|| self.fail(format!("`@inline {callee}`: bad arity or unresolved Const argument")));
-        if n_ret != dsts.len() {
-            self.fail(format!(
-                "`@inline {callee}` returns {n_ret} values, call binds {}",
-                dsts.len()
-            ))
-        };
-        if !(body_inlinable(&body, self.defs)) {
-            self.fail(format!("`@inline {callee}` must be a single tail `return` with only builtin or @inline calls, and no loop/match"))
-        };
-        if self.inline_calls.iter().any(|f| f == callee) {
-            self.fail(format!(
-                "recursive @inline expansion is not supported: {} -> {callee}",
-                self.inline_calls.join(" -> ")
-            ))
-        };
-        // Bind the params from the caller-scope arguments (symbolically where we
-        // can, so a shifted-pointer arg keeps folding into `β`; a `StackBuf` arg
-        // aliases its cell run), then lower the body in a fresh variable
-        // environment, since a function sees only its params. The frame, `one`,
-        // `self_fp`, and range-check bounds stay the caller's: the inlined code
-        // runs in the caller's frame, so they fit.
-        let mut binds: Vec<(String, Binding)> = Vec::new();
-        for (p, a) in params.iter().zip(&rt_args) {
-            let b = if let Some((base, size)) = self.stack_of(a) {
-                Binding::Stack(base, size)
-            } else if let Some(ga) = self.gaddr_of(a) {
-                Binding::Gaddr(ga)
-            } else {
-                Binding::Scalar(self.expr(a))
-            };
-            binds.push((p.clone(), b));
-        }
-        // Only the name bindings reset: the inlined body runs in the caller's
-        // frame, so the caller's `one`, `self_fp`, constant and bound cells all
-        // still name valid cells and stay live.
-        let saved = std::mem::take(&mut self.scope.names);
-        for (p, b) in binds {
-            self.check_not_reserved(&p);
-            self.scope.names.insert(p, Bound { val: b, int: None });
-        }
-        let saved_ret = self.inline_ret.replace(dsts.to_vec());
-        // The body lowers through the CALLER's `FnLower`, so its statements move
-        // `cur_line` into the callee. Restoring it is what keeps the rest of the
-        // caller's expression attributed to the call site rather than to whatever
-        // line the callee happened to end on.
-        let saved_line = self.cur_line;
-        self.inline_calls.push(callee.to_string());
-        for s in &body {
-            self.stmt(s);
-        }
-        let popped = self.inline_calls.pop();
-        debug_assert_eq!(popped.as_deref(), Some(callee));
-        self.inline_ret = saved_ret;
-        self.cur_line = saved_line;
-        self.scope.names = saved;
-        true
-    }
-
-    /// If `callee` declares `Const` parameters, monomorphize: the constant
-    /// arguments (literals, `GEN ** k`, or literal-bound names) substitute into
-    /// a copy of the callee, queued once per distinct constant tuple and named
-    /// `callee__L5_G3`-style, and only the runtime arguments remain.
-    /// A callee's declared return shapes, looked up wherever it lives: an
-    /// ordinary definition sits in `defs`, while a `Const` specialization is
-    /// registered by [`Self::specialize`] in the queue under its mangled name and
-    /// never reaches `defs`. A dispatched `match_range` names specializations, so a
-    /// check that consults only `defs` silently passes on every one of them.
-    /// How many arguments `callee` takes, looked up the same way its return
-    /// shapes are, so a specialization or a generated helper answers too.
-    fn arity_of(&self, callee: &str) -> Option<usize> {
-        self.defs
-            .get(callee)
-            .map(|d| d.params.len())
-            .or_else(|| self.queue.iter().find(|f| f.name == callee).map(|f| f.params.len()))
-    }
-
-    fn return_shapes_of(&self, callee: &str) -> Option<Vec<ReturnShape>> {
-        self.defs.get(callee).map(|d| d.return_shapes.clone()).or_else(|| {
-            self.queue
-                .iter()
-                .find(|f| f.name == callee)
-                .map(|f| f.return_shapes.clone())
-        })
-    }
-
-    fn specialize(&mut self, callee: &str, args: &[Expr]) -> (String, Vec<Expr>) {
-        let defs: &HashMap<String, Func> = self.defs;
-        let Some(def) = defs.get(callee) else {
-            return (callee.to_string(), args.to_vec()); // loop helpers, unknown names
-        };
-        if !def.const_params.contains(&true) {
-            return (callee.to_string(), args.to_vec());
-        }
-        if args.len() != def.params.len() {
-            self.fail(format!("call to `{callee}`: wrong arity"))
-        };
-        let mut tag = String::new();
-        let (mut rt_params, mut rt_args, mut substs) = (Vec::new(), Vec::new(), Vec::new());
-        for ((p, &is_const), a) in def.params.iter().zip(&def.const_params).zip(args) {
-            if !is_const {
-                rt_params.push(p.clone());
-                rt_args.push(a.clone());
-                continue;
-            }
-            let c = self.const_arg(a).unwrap_or_else(|| {
-                self.fail(format!(
-                    "argument for Const parameter `{p}` of `{callee}` must be a compile-time \
-                     constant, got `{a:?}`"
-                ))
-            });
-            tag.push_str(&match &c {
-                Expr::Lit(n) => format!("_L{n}"),
-                Expr::GPow(k) => format!("_G{k}"),
-                _ => unreachable!(),
-            });
-            substs.push((p.clone(), c));
-        }
-        let name = format!("{callee}_{tag}");
-        if !self.queue.iter().any(|f| f.name == name) {
-            if self.queue.len() >= 10_000 {
-                self.fail("Const specialization explosion (recursive constants?)")
-            };
-            let mut body = def.body.clone();
-            for (p, c) in &substs {
-                body = subst_stmts(&body, p, c);
-            }
-            let const_params = vec![false; rt_params.len()];
-            self.queue.push(Func {
-                name: name.clone(),
-                params: rt_params,
-                const_params,
-                n_ret: def.n_ret,
-                return_shapes: def.return_shapes.clone(),
-                body,
-                inline: false,
-            });
-        }
-        (name, rt_args)
-    }
-
-    /// Lower a call. Return values land in `dsts_in` when given (write-once, so
-    /// distinct arms of a `match_range` may share the same cells), else in fresh
-    /// cells, sparing the caller a temp-then-copy.
-    fn lower_call(
-        &mut self,
-        callee: &str,
-        args: &[Expr],
-        n_ret: usize,
-        cond: Option<Off>,
-        dsts_in: Option<&[Off]>,
-        tail: bool,
-    ) -> Vec<Off> {
-        // Every parameter must be supplied. A missing argument leaves the
-        // callee's argument cell unwritten, hence prover-chosen, so an `assert`
-        // reading it is vacuous; a surplus one lands on the callee's first
-        // return slot, because caller and callee place the return area from
-        // their own idea of the argument count. Only `specialize` checked this,
-        // and only for a callee declaring `Const` parameters.
-        match self.arity_of(callee) {
-            Some(want) if want != args.len() => {
-                let plural = if want == 1 { "argument" } else { "arguments" };
-                self.fail(format!("`{callee}` takes {want} {plural}, got {}", args.len()))
-            }
-            // Nothing by that name is going to be lowered, so the entry pc it
-            // needs will not exist. Caught here, where there is a line: a typo, a
-            // statement-only builtin used as a value (`x = assert_in_k(a, b)`),
-            // or an `@inline` callee reached where inlining did not happen, all
-            // used to die later in `resolve` as a bare `no entry found for key`.
-            None => self.fail(format!(
-                "no function named `{callee}`. A builtin that writes into a destination \
-                 (`blake2s`, `assert_in_k`, a `hint_*`) is a statement and returns nothing, and an \
-                 `@inline` function is expanded at its call site rather than called"
-            )),
-            _ => {}
-        }
-        let (callee, args) = self.specialize(callee, args);
-        let (callee, args) = (callee.as_str(), args.as_slice());
-        let arg_offs: Vec<Off> = args.iter().map(|a| self.expr(a)).collect();
-        let nfp = self.fresh();
-        let entry = self.fresh();
-        // Resolve the jump condition up front: `self.one()` may emit a `SET`, and
-        // nothing may sit between the retpc `DEREF` and the `JUMP` (the `g²·pc`
-        // return target assumes the `JUMP` is exactly one instruction later).
-        let oc = cond.unwrap_or_else(|| self.one());
-        self.set(entry, KVal::Entry(callee.to_string()));
-
-        // The frame-pointer hint fires before the first DEREF that reads `nfp`.
-        self.pending.push(Hint::AllocFrame {
-            ptr: nfp,
-            callee: callee.to_string(),
-        });
-        for (i, &ao) in arg_offs.iter().enumerate() {
-            self.deref(nfp, Abi::arg(i as u32), ao, DerefMode::Cell);
-        }
-        if tail {
-            // Tail call: hand the callee OUR return target, so it returns to our
-            // caller and we are never resumed. Cells 0/1 of this frame already
-            // hold that target (written by whoever called us).
-            self.deref(nfp, 1, 1, DerefMode::Cell); // retfp := our retfp
-            self.deref(nfp, 0, 0, DerefMode::Cell); // retpc := our retpc
-        } else {
-            self.deref(nfp, 1, 0, DerefMode::Fp); // retfp
-            self.deref(nfp, 0, 0, DerefMode::Pc); // retpc = g²·pc
-        }
-        self.emit(LOp::Jump { oc, od: entry, of: nfp });
-
-        let n_args = args.len() as u32;
-        let dsts: Vec<Off> = match dsts_in {
-            Some(d) => d.to_vec(),
-            None => (0..n_ret).map(|_| self.fresh()).collect(),
-        };
-        for (i, &d) in dsts.iter().enumerate() {
-            self.deref(nfp, Abi::ret(n_args, i as u32), d, DerefMode::Cell);
-        }
-        dsts
     }
 
     fn stmt(&mut self, s: &Stmt) {
@@ -2360,229 +1416,6 @@ impl FnLower<'_> {
         }
     }
 
-    /// The statement-position builtins, `true` if `f` was one of them (else the
-    /// caller emits an ordinary call). The `hint_*` ones queue prover-side
-    /// advice, re-checked in-circuit by their caller: `hint_decompose_bits`
-    /// writes a value's bits into a buffer, `hint_decompose_bits_exponent` the
-    /// bits of `n` where the value is `g^n` (a bounded dlog at witness
-    /// generation), `hint_f192_limbs` a value's coordinate limbs.
-    fn lower_builtin(&mut self, f: &str, args: &[Expr]) -> bool {
-        match f {
-            "hint_decompose_bits" | "hint_decompose_bits_exponent" => {
-                if args.len() != 3 {
-                    self.fail(format!(
-                        "{f} takes three arguments, `(bits, value, nbits)`, got {}",
-                        args.len()
-                    ))
-                };
-                let nbits = self.const_index(&args[2]);
-                let bits = self.bits_dest(&args[0], nbits, f);
-                let value = self.expr(&args[1]);
-                self.pending.push(Hint::Resolved(if f == "hint_decompose_bits" {
-                    RHint::BitDecompose { value, bits, nbits }
-                } else {
-                    RHint::BitDecomposeExp { value, bits, nbits }
-                }));
-            }
-            "blake2s" => self.lower_blake2s(args),
-            "assert_in_k" => {
-                if args.len() != 2 {
-                    self.fail("assert_in_k(a, b) takes two scalar cells")
-                };
-                let a = self.expr(&args[0]);
-                let b = self.expr(&args[1]);
-                let zero = self.zero();
-                self.emit(LOp::Jump { oc: zero, od: a, of: b });
-            }
-            "hint_f192_limbs" => {
-                if args.len() != 2 {
-                    self.fail(format!(
-                        "hint_f192_limbs takes two arguments, `(dest, value)`, got {}",
-                        args.len()
-                    ))
-                };
-                let (base, len) = self
-                    .stack_of(&args[0])
-                    .expect("hint_f192_limbs destination must be a StackBuf");
-                if !((1..=3).contains(&len)) {
-                    self.fail("hint_f192_limbs destination must have 1..=3 cells")
-                };
-                let value = self.expr(&args[1]);
-                let value = self.word_src(value);
-                // Names the physical cells, as the two consumers above do, so the run
-                // has to hold real values before the hint fills it. The common
-                // destination is a list literal (`limbs = [0, 0, 0]`), whose every
-                // element goes through `stack_store` and so defers.
-                self.materialize_run(base, len);
-                self.pending
-                    .push(Hint::Resolved(RHint::FieldLimbs { value, base, len }));
-            }
-            _ => return false,
-        }
-        true
-    }
-
-    /// `blake2s(a, b, out)`: the digest of the two 256-bit operands lands in the
-    /// existing 2-cell run `out` (write-once: if `out` was already written, this
-    /// asserts the digest equals it). A heap `out` slice takes the digest via a
-    /// fresh stack pair and two `DEREF`s after the hash, the store direction
-    /// being the same instruction as the load (write-once fills the unset side).
-    /// Keyword arguments set the compile-time metadata.
-    fn lower_blake2s(&mut self, args: &[Expr]) {
-        let first_kw = args
-            .iter()
-            .position(|a| matches!(a, Expr::Call(name, _) if name.starts_with("__kw_")))
-            .unwrap_or(args.len());
-        if first_kw != 3 {
-            self.fail("blake2s takes three positional arguments: (a, b, out)")
-        };
-        if !(args[first_kw..]
-            .iter()
-            .all(|a| matches!(a, Expr::Call(name, v) if name.starts_with("__kw_") && v.len() == 1)))
-        {
-            self.fail("keyword arguments must follow the three positional blake2s arguments")
-        };
-        let mut kwargs: HashMap<&str, &Expr> = HashMap::new();
-        for kw in &args[first_kw..] {
-            let Expr::Call(name, value) = kw else { unreachable!() };
-            let key = name.strip_prefix("__kw_").unwrap();
-            if kwargs.insert(key, &value[0]).is_some() {
-                self.fail(format!("duplicate blake2s keyword `{key}`"))
-            };
-        }
-        let allowed = ["cv", "counter", "final", "last_node"];
-        if !(kwargs.keys().all(|k| allowed.contains(k))) {
-            let bad: Vec<&&str> = kwargs.keys().filter(|k| !allowed.contains(k)).collect();
-            self.fail(format!(
-                "unknown blake2s keyword {bad:?}; the keywords are {allowed:?}"
-            ))
-        };
-        let customized = kwargs.keys().any(|k| matches!(*k, "counter" | "final" | "last_node"));
-        if kwargs.contains_key("cv") && !customized {
-            self.fail("blake2s with cv= requires counter=, since a chained block is not the default one-block hash")
-        };
-
-        let a = self.blake2s_input(&args[0]);
-        let b = self.blake2s_input(&args[1]);
-        let (c, heap_out) = match self.blake2s_operand(&args[2]) {
-            CellRun::Stack { base, .. } => {
-                self.materialize_run(base, 2);
-                (base, None)
-            }
-            CellRun::Heap { ptr, lo, .. } => (self.alloc_stack(2), Some((ptr, lo))),
-        };
-        let cv = if let Some(value) = kwargs.get("cv") {
-            self.blake2s_cv(value)
-        } else {
-            self.default_blake2s_cv()
-        };
-        let const_kw = |this: &Self, name: &str, default: u128| -> u128 {
-            kwargs
-                .get(name)
-                .map(|e| {
-                    this.try_const_int(e).unwrap_or_else(|| {
-                        self.fail(format!("BLAKE2s `{name}` must be a compile-time integer, got `{e:?}`"))
-                    })
-                })
-                .unwrap_or(default)
-        };
-        // BLAKE2s metadata is just the cumulative byte counter and two flags, so
-        // a multi-block hash is `counter = 64 * blocks_before + bytes_in_this_block`
-        // and `final = 1` on the last block. The default is the one-block hash of
-        // a full 64-byte input, which is what `vmhash::compress` and every Merkle
-        // node use.
-        let counter = u64::try_from(const_kw(self, "counter", 64)).expect("BLAKE2s counter does not fit in u64");
-        let f0 = if const_kw(self, "final", if customized { 0 } else { 1 }) != 0 {
-            lean_vm::hash_flock::FINAL_FLAG
-        } else {
-            0
-        };
-        let f1 = if const_kw(self, "last_node", 0) != 0 {
-            u32::MAX
-        } else {
-            0
-        };
-        let metadata = lean_vm::hash_flock::metadata(counter, f0, f1);
-        // Each operand is two 128-bit chunk cells; the flexible opcode addresses
-        // the four input cells independently (`blake2s_input` forwards the real
-        // chunk sources where it can). The digest occupies the two consecutive
-        // output cells `c, g·c`.
-        self.emit(LOp::Blake2s {
-            ins: [a[0], a[1], b[0], b[1]],
-            cv,
-            c,
-            metadata,
-        });
-        if let Some((ptr, lo)) = heap_out {
-            for k in 0..2 {
-                self.deref(ptr, lo + k, c + k, DerefMode::Cell);
-            }
-        }
-    }
-
-    fn lower_return(&mut self, exprs: &[Expr]) {
-        // Inlined (`@inline`): bind the return values into the caller's cells
-        // and fall through: this is the body's tail return, so no jump is needed.
-        if let Some(dsts) = self.inline_ret.clone() {
-            // Each returned value is bound into the caller independently, exactly
-            // as a `let name = <that expr>` would: a `StackBuf` or a folded
-            // g-address hands over its run/pointer (alias, not copies: allocated
-            // in the caller's frame, so it outlives the inline scope), a scalar is
-            // copied into its dst cell. The per-slot record lets the caller's
-            // `let`/tuple pick the right binding, so a fused
-            // `fs, x, cur = fs_next(fs, cur)` returns a StackBuf, a scalar, and an
-            // advanced cursor together.
-            let mut binds = Vec::with_capacity(dsts.len());
-            for (e, &d) in exprs.iter().zip(&dsts) {
-                binds.push(if let Some((base, size)) = self.stack_of(e) {
-                    RetBind::Stack(base, size)
-                } else if let Some(ga) = self.gaddr_of(e) {
-                    RetBind::Gaddr(ga)
-                } else {
-                    self.expr_into(e, d);
-                    RetBind::Scalar
-                });
-            }
-            self.inline_stack_ret = Some(binds);
-            return;
-        }
-        if self.is_main {
-            return; // a `return` in main is a no-op; main halts via the trailing sentinel jump (lower_func).
-        }
-        let ret_base = Abi::ret(self.n_args, 0);
-        if exprs.len() != self.return_shapes.len() {
-            self.fail(format!(
-                "function returns {} values here, but its ABI declares {}",
-                exprs.len(),
-                self.return_shapes.len()
-            ))
-        };
-        // Each logical value lands straight in its flattened return area. A
-        // StackBuf is copied cell-by-cell because its callee-frame offsets are
-        // not meaningful after control returns to the caller.
-        let mut ret = ret_base;
-        for (e, shape) in exprs.iter().zip(self.return_shapes.clone()) {
-            match shape {
-                ReturnShape::Scalar => self.expr_into(e, ret),
-                ReturnShape::StackBuf(size) => {
-                    let (base, actual) = self
-                        .stack_of(e)
-                        .unwrap_or_else(|| self.fail(format!("expected a StackBuf({size}) return, got `{e:?}`")));
-                    if actual != size {
-                        self.fail(format!("returned StackBuf has size {actual}, expected {size}"))
-                    };
-                    for k in 0..size {
-                        let src = self.word_src(base + k);
-                        self.copy(src, ret + k);
-                    }
-                }
-            }
-            ret += shape.cells();
-        }
-        let one = self.one();
-        self.emit(LOp::Jump { oc: one, od: 0, of: 1 });
-    }
-
     /// `for i in mul_range(GEN**lo, GEN**hi)` → a single tail-recursive helper, with the
     /// exit test folded into the recursion's condition (no separate branch, no
     /// is-zero gadget):
@@ -2739,37 +1572,6 @@ impl FnLower<'_> {
                 self.stmt(&stmt);
             }
         }
-    }
-}
-
-/// A body safe to inline: a single **tail** `return`, and no construct whose
-/// lowering needs its own frame or a dispatch: a non-inline user call, a
-/// runtime loop, or a match (which would reload a frame pointer that is no
-/// longer the callee's). Builtins and nested `@inline` calls are fine;
-/// `unroll`/`if` are compile-time / same-frame and recurse into.
-fn body_inlinable(body: &[Stmt], defs: &HashMap<String, Func>) -> bool {
-    matches!(body.split_last(), Some((last, rest)) if matches!(last.kind, StmtKind::Return(_))
-        && rest.iter().all(|s| stmt_inline_safe(s, defs)))
-}
-
-fn stmt_inline_safe(s: &Stmt, defs: &HashMap<String, Func>) -> bool {
-    match &s.kind {
-        StmtKind::Let(..)
-        | StmtKind::Store(..)
-        | StmtKind::HintWitness { .. }
-        | StmtKind::Print { .. }
-        | StmtKind::AssertEq(..)
-        | StmtKind::AssertNe(..)
-        | StmtKind::AssertLt(..) => true,
-        StmtKind::Call(f, _) => {
-            f == "blake2s" || f == "assert_in_k" || f == "hint_f192_limbs" || defs.get(f).is_some_and(|d| d.inline)
-        }
-        StmtKind::If { then, els, .. } => {
-            then.iter().all(|s| stmt_inline_safe(s, defs)) && els.iter().all(|s| stmt_inline_safe(s, defs))
-        }
-        StmtKind::Unroll { body, .. } => body.iter().all(|s| stmt_inline_safe(s, defs)),
-        // Return (non-tail), For, Match, LetMatchRange, LetTuple, CallIfNe, user Call.
-        _ => false,
     }
 }
