+++ crates/lean_compiler/src/lower/call.rs
@@ -1,0 +1,561 @@
+//! The call boundary: arguments in, return values out, and the two ways a
+//! callee can disappear into its caller.
+//!
+//! Every frame here is laid out by [`Abi`], and the rule that makes a call sound
+//! is that CALLER AND CALLEE MUST AGREE ON THE ARITY. They place the return area
+//! from their own idea of the argument count, so a mismatch does not merely lose
+//! a value: a missing argument leaves the callee's cell unwritten and therefore
+//! prover-chosen, and makes the caller bind a callee parameter cell as a return;
+//! a surplus one overwrites the callee's first return slot. Two of the three
+//! soundness bugs found in an audit of this crate were that check missing, once
+//! on the ordinary path and once on the fused `match_range` dispatch.
+//!
+//! A callee vanishes into its caller two ways. `Const` specialization
+//! monomorphises it per constant tuple and lowers the copy; `@inline` expands
+//! the body into the caller's own frame, which is why the caller's `one`,
+//! `self_fp` and constant cells stay valid across it and only the name bindings
+//! reset.
+
+use super::*;
+
+/// How an inlined tail return binds in the caller: a `StackBuf` run and a folded
+/// g-address alias at zero copies, while anything else (a plain scalar, or a real
+/// call, which records no [`RetBind`]) takes the destination cell it wrote.
+pub(super) fn ret_binding(b: Option<RetBind>, dst: Off) -> Binding {
+    match b {
+        Some(RetBind::Stack(base, size)) => Binding::Stack(base, size),
+        Some(RetBind::Gaddr(ga)) => Binding::Gaddr(ga),
+        _ => Binding::Scalar(dst),
+    }
+}
+
+/// A body safe to inline: a single **tail** `return`, and no construct whose
+/// lowering needs its own frame or a dispatch: a non-inline user call, a
+/// runtime loop, or a match (which would reload a frame pointer that is no
+/// longer the callee's). Builtins and nested `@inline` calls are fine;
+/// `unroll`/`if` are compile-time / same-frame and recurse into.
+pub(super) fn body_inlinable(body: &[Stmt], defs: &HashMap<String, Func>) -> bool {
+    matches!(body.split_last(), Some((last, rest)) if matches!(last.kind, StmtKind::Return(_))
+        && rest.iter().all(|s| stmt_inline_safe(s, defs)))
+}
+
+pub(super) fn stmt_inline_safe(s: &Stmt, defs: &HashMap<String, Func>) -> bool {
+    match &s.kind {
+        StmtKind::Let(..)
+        | StmtKind::Store(..)
+        | StmtKind::HintWitness { .. }
+        | StmtKind::Print { .. }
+        | StmtKind::AssertEq(..)
+        | StmtKind::AssertNe(..)
+        | StmtKind::AssertLt(..) => true,
+        StmtKind::Call(f, _) => {
+            f == "blake2s" || f == "assert_in_k" || f == "hint_f192_limbs" || defs.get(f).is_some_and(|d| d.inline)
+        }
+        StmtKind::If { then, els, .. } => {
+            then.iter().all(|s| stmt_inline_safe(s, defs)) && els.iter().all(|s| stmt_inline_safe(s, defs))
+        }
+        StmtKind::Unroll { body, .. } => body.iter().all(|s| stmt_inline_safe(s, defs)),
+        // Return (non-tail), For, Match, LetMatchRange, LetTuple, CallIfNe, user Call.
+        _ => false,
+    }
+}
+
+impl FnLower<'_> {
+    /// Lower a call. Return values land in `dsts_in` when given (write-once, so
+    /// distinct arms of a `match_range` may share the same cells), else in fresh
+    /// cells, sparing the caller a temp-then-copy.
+    pub(super) fn lower_call(
+        &mut self,
+        callee: &str,
+        args: &[Expr],
+        n_ret: usize,
+        cond: Option<Off>,
+        dsts_in: Option<&[Off]>,
+        tail: bool,
+    ) -> Vec<Off> {
+        // Every parameter must be supplied. A missing argument leaves the
+        // callee's argument cell unwritten, hence prover-chosen, so an `assert`
+        // reading it is vacuous; a surplus one lands on the callee's first
+        // return slot, because caller and callee place the return area from
+        // their own idea of the argument count. Only `specialize` checked this,
+        // and only for a callee declaring `Const` parameters.
+        match self.arity_of(callee) {
+            Some(want) if want != args.len() => {
+                let plural = if want == 1 { "argument" } else { "arguments" };
+                self.fail(format!("`{callee}` takes {want} {plural}, got {}", args.len()))
+            }
+            // Nothing by that name is going to be lowered, so the entry pc it
+            // needs will not exist. Caught here, where there is a line: a typo, a
+            // statement-only builtin used as a value (`x = assert_in_k(a, b)`),
+            // or an `@inline` callee reached where inlining did not happen, all
+            // used to die later in `resolve` as a bare `no entry found for key`.
+            None => self.fail(format!(
+                "no function named `{callee}`. A builtin that writes into a destination \
+                 (`blake2s`, `assert_in_k`, a `hint_*`) is a statement and returns nothing, and an \
+                 `@inline` function is expanded at its call site rather than called"
+            )),
+            _ => {}
+        }
+        let (callee, args) = self.specialize(callee, args);
+        let (callee, args) = (callee.as_str(), args.as_slice());
+        let arg_offs: Vec<Off> = args.iter().map(|a| self.expr(a)).collect();
+        let nfp = self.fresh();
+        let entry = self.fresh();
+        // Resolve the jump condition up front: `self.one()` may emit a `SET`, and
+        // nothing may sit between the retpc `DEREF` and the `JUMP` (the `g²·pc`
+        // return target assumes the `JUMP` is exactly one instruction later).
+        let oc = cond.unwrap_or_else(|| self.one());
+        self.set(entry, KVal::Entry(callee.to_string()));
+
+        // The frame-pointer hint fires before the first DEREF that reads `nfp`.
+        self.pending.push(Hint::AllocFrame {
+            ptr: nfp,
+            callee: callee.to_string(),
+        });
+        for (i, &ao) in arg_offs.iter().enumerate() {
+            self.deref(nfp, Abi::arg(i as u32), ao, DerefMode::Cell);
+        }
+        if tail {
+            // Tail call: hand the callee OUR return target, so it returns to our
+            // caller and we are never resumed. Cells 0/1 of this frame already
+            // hold that target (written by whoever called us).
+            self.deref(nfp, 1, 1, DerefMode::Cell); // retfp := our retfp
+            self.deref(nfp, 0, 0, DerefMode::Cell); // retpc := our retpc
+        } else {
+            self.deref(nfp, 1, 0, DerefMode::Fp); // retfp
+            self.deref(nfp, 0, 0, DerefMode::Pc); // retpc = g²·pc
+        }
+        self.emit(LOp::Jump { oc, od: entry, of: nfp });
+
+        let n_args = args.len() as u32;
+        let dsts: Vec<Off> = match dsts_in {
+            Some(d) => d.to_vec(),
+            None => (0..n_ret).map(|_| self.fresh()).collect(),
+        };
+        for (i, &d) in dsts.iter().enumerate() {
+            self.deref(nfp, Abi::ret(n_args, i as u32), d, DerefMode::Cell);
+        }
+        dsts
+    }
+
+    /// `names = match_range(log(x), …, lambda k: f(args, k))` fused: the arms all
+    /// call one of `callees` (specializations sharing the arg/return layout) with
+    /// the same runtime `args`, so build the callee frame **once** and let the
+    /// dispatch jump straight into the selected entry, which returns to the join.
+    /// Each taken arm is then just the trampoline's `SET entry; JUMP`: no
+    /// per-arm frame setup, call, or return jump.
+    pub(super) fn lower_dispatched_call(&mut self, names: &[String], x: &Expr, callees: &[String], rt_args: &[Expr]) {
+        let n_args = rt_args.len() as u32;
+        // The join below reads one return cell per bound name, so every callee has
+        // to declare exactly that many. Unchecked, a name past a callee's arity
+        // `DEREF`s a frame offset nothing on that path writes, and since the shared
+        // frame is sized to the LARGEST callee the offset exists: the surplus name
+        // binds a prover-chosen word. The non-fused path enforces this
+        // ([`Self::call_into`]), so leaving it out here means one source is rejected
+        // by one lowering of `match_range` and silently miscompiled by the other.
+        for callee in callees {
+            // Arguments for the same reason as returns below: the shared frame
+            // is sized to the largest callee, so a callee expecting more than
+            // the arms supply reads a cell that exists and nothing writes.
+            if let Some(want) = self.arity_of(callee)
+                && want != rt_args.len()
+            {
+                let plural = if want == 1 { "argument" } else { "arguments" };
+                self.fail(format!(
+                    "`{callee}` takes {want} {plural}, dispatched call passes {}",
+                    rt_args.len()
+                ))
+            }
+            let Some(shapes) = self.return_shapes_of(callee) else {
+                continue;
+            };
+            if shapes.len() != names.len() {
+                self.fail(format!(
+                    "`{callee}` returns {} values, dispatched call binds {}",
+                    shapes.len(),
+                    names.len()
+                ))
+            };
+            if shapes.iter().any(|s| *s != ReturnShape::Scalar) {
+                self.fail(format!(
+                    "`{callee}`: a multi-cell StackBuf return cannot cross a dispatched join"
+                ))
+            };
+        }
+        let rcells: Vec<Off> = names.iter().map(|_| self.fresh()).collect();
+
+        // Shared callee frame: args, retfp, and retpc = the join (so the callee
+        // returns straight past the dispatch). Evaluated once.
+        let arg_offs: Vec<Off> = rt_args.iter().map(|a| self.expr(a)).collect();
+        let xo = self.expr(x);
+        let one = self.one();
+        let sfp = self.self_fp();
+
+        let nfp = self.fresh();
+        self.pending.push(Hint::AllocFrameMax {
+            ptr: nfp,
+            callees: callees.to_vec(),
+        });
+        for (i, &ao) in arg_offs.iter().enumerate() {
+            self.deref(nfp, Abi::arg(i as u32), ao, DerefMode::Cell);
+        }
+        self.deref(nfp, Abi::RET_FP, 0, DerefMode::Fp);
+        let join_cell = self.fresh();
+        let join_set = self.code.len();
+        self.set(join_cell, KVal::Local(0)); // patched: the join pc
+        self.deref(nfp, Abi::RET_PC, join_cell, DerefMode::Cell); // retpc = join
+
+        let kset = self.emit_dispatch(xo, one, sfp);
+
+        // Trampoline: slot j enters `callees[j]` with fp = nfp; the callee's own
+        // `return` jumps to retpc (the join) in the caller frame.
+        self.patch_local(kset, self.code.len());
+        self.emit_slots(callees.len(), one, nfp, |j| KVal::Entry(callees[j].clone()));
+
+        // Join: read the return values (written by whichever callee ran).
+        self.patch_local(join_set, self.code.len());
+        for (i, &r) in rcells.iter().enumerate() {
+            self.deref(nfp, Abi::ret(n_args, i as u32), r, DerefMode::Cell);
+        }
+
+        self.bind_join(names, &rcells);
+    }
+
+    /// Inline an `@inline` `callee(args)` into the current frame, binding its
+    /// return values straight into `dsts`: no frame setup, no argument/return
+    /// plumbing, no call/return jumps. Returns `false` for a non-`@inline`
+    /// callee (the caller emits a real call). Panics if an `@inline` function
+    /// isn't inlinable ([`body_inlinable`]) or its `Const` args don't resolve.
+    pub(super) fn try_inline(&mut self, callee: &str, args: &[Expr], dsts: &[Off]) -> bool {
+        if !self.defs.get(callee).is_some_and(|d| d.inline) {
+            return false;
+        }
+        let (params, rt_args, body, n_ret) = self
+            .specialized_body(callee, args)
+            .unwrap_or_else(|| self.fail(format!("`@inline {callee}`: bad arity or unresolved Const argument")));
+        if n_ret != dsts.len() {
+            self.fail(format!(
+                "`@inline {callee}` returns {n_ret} values, call binds {}",
+                dsts.len()
+            ))
+        };
+        if !(body_inlinable(&body, self.defs)) {
+            self.fail(format!("`@inline {callee}` must be a single tail `return` with only builtin or @inline calls, and no loop/match"))
+        };
+        if self.inline_calls.iter().any(|f| f == callee) {
+            self.fail(format!(
+                "recursive @inline expansion is not supported: {} -> {callee}",
+                self.inline_calls.join(" -> ")
+            ))
+        };
+        // Bind the params from the caller-scope arguments (symbolically where we
+        // can, so a shifted-pointer arg keeps folding into `β`; a `StackBuf` arg
+        // aliases its cell run), then lower the body in a fresh variable
+        // environment, since a function sees only its params. The frame, `one`,
+        // `self_fp`, and range-check bounds stay the caller's: the inlined code
+        // runs in the caller's frame, so they fit.
+        let mut binds: Vec<(String, Binding)> = Vec::new();
+        for (p, a) in params.iter().zip(&rt_args) {
+            let b = if let Some((base, size)) = self.stack_of(a) {
+                Binding::Stack(base, size)
+            } else if let Some(ga) = self.gaddr_of(a) {
+                Binding::Gaddr(ga)
+            } else {
+                Binding::Scalar(self.expr(a))
+            };
+            binds.push((p.clone(), b));
+        }
+        // Only the name bindings reset: the inlined body runs in the caller's
+        // frame, so the caller's `one`, `self_fp`, constant and bound cells all
+        // still name valid cells and stay live.
+        let saved = std::mem::take(&mut self.scope.names);
+        for (p, b) in binds {
+            self.check_not_reserved(&p);
+            self.scope.names.insert(p, Bound { val: b, int: None });
+        }
+        let saved_ret = self.inline_ret.replace(dsts.to_vec());
+        // The body lowers through the CALLER's `FnLower`, so its statements move
+        // `cur_line` into the callee. Restoring it is what keeps the rest of the
+        // caller's expression attributed to the call site rather than to whatever
+        // line the callee happened to end on.
+        let saved_line = self.cur_line;
+        self.inline_calls.push(callee.to_string());
+        for s in &body {
+            self.stmt(s);
+        }
+        let popped = self.inline_calls.pop();
+        debug_assert_eq!(popped.as_deref(), Some(callee));
+        self.inline_ret = saved_ret;
+        self.cur_line = saved_line;
+        self.scope.names = saved;
+        true
+    }
+
+    pub(super) fn lower_return(&mut self, exprs: &[Expr]) {
+        // Inlined (`@inline`): bind the return values into the caller's cells
+        // and fall through: this is the body's tail return, so no jump is needed.
+        if let Some(dsts) = self.inline_ret.clone() {
+            // Each returned value is bound into the caller independently, exactly
+            // as a `let name = <that expr>` would: a `StackBuf` or a folded
+            // g-address hands over its run/pointer (alias, not copies: allocated
+            // in the caller's frame, so it outlives the inline scope), a scalar is
+            // copied into its dst cell. The per-slot record lets the caller's
+            // `let`/tuple pick the right binding, so a fused
+            // `fs, x, cur = fs_next(fs, cur)` returns a StackBuf, a scalar, and an
+            // advanced cursor together.
+            let mut binds = Vec::with_capacity(dsts.len());
+            for (e, &d) in exprs.iter().zip(&dsts) {
+                binds.push(if let Some((base, size)) = self.stack_of(e) {
+                    RetBind::Stack(base, size)
+                } else if let Some(ga) = self.gaddr_of(e) {
+                    RetBind::Gaddr(ga)
+                } else {
+                    self.expr_into(e, d);
+                    RetBind::Scalar
+                });
+            }
+            self.inline_stack_ret = Some(binds);
+            return;
+        }
+        if self.is_main {
+            return; // a `return` in main is a no-op; main halts via the trailing sentinel jump (lower_func).
+        }
+        let ret_base = Abi::ret(self.n_args, 0);
+        if exprs.len() != self.return_shapes.len() {
+            self.fail(format!(
+                "function returns {} values here, but its ABI declares {}",
+                exprs.len(),
+                self.return_shapes.len()
+            ))
+        };
+        // Each logical value lands straight in its flattened return area. A
+        // StackBuf is copied cell-by-cell because its callee-frame offsets are
+        // not meaningful after control returns to the caller.
+        let mut ret = ret_base;
+        for (e, shape) in exprs.iter().zip(self.return_shapes.clone()) {
+            match shape {
+                ReturnShape::Scalar => self.expr_into(e, ret),
+                ReturnShape::StackBuf(size) => {
+                    let (base, actual) = self
+                        .stack_of(e)
+                        .unwrap_or_else(|| self.fail(format!("expected a StackBuf({size}) return, got `{e:?}`")));
+                    if actual != size {
+                        self.fail(format!("returned StackBuf has size {actual}, expected {size}"))
+                    };
+                    for k in 0..size {
+                        let src = self.word_src(base + k);
+                        self.copy(src, ret + k);
+                    }
+                }
+            }
+            ret += shape.cells();
+        }
+        let one = self.one();
+        self.emit(LOp::Jump { oc: one, od: 0, of: 1 });
+    }
+
+    pub(super) fn specialize(&mut self, callee: &str, args: &[Expr]) -> (String, Vec<Expr>) {
+        let defs: &HashMap<String, Func> = self.defs;
+        let Some(def) = defs.get(callee) else {
+            return (callee.to_string(), args.to_vec()); // loop helpers, unknown names
+        };
+        if !def.const_params.contains(&true) {
+            return (callee.to_string(), args.to_vec());
+        }
+        if args.len() != def.params.len() {
+            self.fail(format!("call to `{callee}`: wrong arity"))
+        };
+        let mut tag = String::new();
+        let (mut rt_params, mut rt_args, mut substs) = (Vec::new(), Vec::new(), Vec::new());
+        for ((p, &is_const), a) in def.params.iter().zip(&def.const_params).zip(args) {
+            if !is_const {
+                rt_params.push(p.clone());
+                rt_args.push(a.clone());
+                continue;
+            }
+            let c = self.const_arg(a).unwrap_or_else(|| {
+                self.fail(format!(
+                    "argument for Const parameter `{p}` of `{callee}` must be a compile-time \
+                     constant, got `{a:?}`"
+                ))
+            });
+            tag.push_str(&match &c {
+                Expr::Lit(n) => format!("_L{n}"),
+                Expr::GPow(k) => format!("_G{k}"),
+                _ => unreachable!(),
+            });
+            substs.push((p.clone(), c));
+        }
+        let name = format!("{callee}_{tag}");
+        if !self.queue.iter().any(|f| f.name == name) {
+            if self.queue.len() >= 10_000 {
+                self.fail("Const specialization explosion (recursive constants?)")
+            };
+            let mut body = def.body.clone();
+            for (p, c) in &substs {
+                body = subst_stmts(&body, p, c);
+            }
+            let const_params = vec![false; rt_params.len()];
+            self.queue.push(Func {
+                name: name.clone(),
+                params: rt_params,
+                const_params,
+                n_ret: def.n_ret,
+                return_shapes: def.return_shapes.clone(),
+                body,
+                inline: false,
+            });
+        }
+        (name, rt_args)
+    }
+
+    /// Lower a call; returns one caller offset per source-level return value.
+    /// A real-call StackBuf return is flattened into consecutive ABI cells and
+    /// copied into a fresh consecutive run in the caller. `inline_stack_ret`
+    /// describes those logical bindings to the surrounding let/tuple lowering.
+    pub(super) fn call(&mut self, callee: &str, args: &[Expr], n_ret: usize) -> Vec<Off> {
+        if callee == "blake2s" {
+            self.fail("blake2s is a statement: `blake2s(a, b, out)` writes the digest into the 2-cell stack run `out`")
+        };
+        self.inline_stack_ret = None;
+        if self.defs.get(callee).is_some_and(|d| d.inline) {
+            let dsts: Vec<Off> = (0..n_ret).map(|_| self.fresh()).collect();
+            self.call_into(callee, args, &dsts);
+            return dsts;
+        }
+
+        let shapes = self
+            .defs
+            .get(callee)
+            .map(|d| d.return_shapes.clone())
+            .unwrap_or_else(|| vec![ReturnShape::Scalar; n_ret]);
+        if shapes.len() != n_ret {
+            self.fail(format!(
+                "`{callee}` returns {} values, call binds {n_ret}",
+                shapes.len()
+            ))
+        };
+        let mut logical = Vec::with_capacity(n_ret);
+        let mut physical = Vec::new();
+        let mut binds = Vec::with_capacity(n_ret);
+        for shape in shapes {
+            match shape {
+                ReturnShape::Scalar => {
+                    let dst = self.fresh();
+                    logical.push(dst);
+                    physical.push(dst);
+                    binds.push(RetBind::Scalar);
+                }
+                ReturnShape::StackBuf(size) => {
+                    if size == 0 {
+                        self.fail("a returned StackBuf must not be empty")
+                    };
+                    let base = self.alloc_stack(size);
+                    logical.push(base);
+                    physical.extend(base..base + size);
+                    binds.push(RetBind::Stack(base, size));
+                }
+            }
+        }
+        self.lower_call(callee, args, physical.len(), None, Some(&physical), false);
+        self.inline_stack_ret = Some(binds);
+        logical
+    }
+
+    /// Evaluate `callee(args)` into `dsts`, inlining the callee when it is
+    /// `@inline` ([`Self::try_inline`]), else a real call.
+    pub(super) fn call_into(&mut self, callee: &str, args: &[Expr], dsts: &[Off]) {
+        if callee == "blake2s" {
+            self.fail("blake2s is a statement, not a value-returning call")
+        };
+        if !self.try_inline(callee, args, dsts) {
+            if let Some(def) = self.defs.get(callee) {
+                if def.return_shapes.len() != dsts.len() {
+                    self.fail(format!(
+                        "`{callee}` returns {} values, call binds {}",
+                        def.return_shapes.len(),
+                        dsts.len()
+                    ))
+                };
+                if def.return_shapes.iter().any(|s| *s != ReturnShape::Scalar) {
+                    self.fail("a normal function's multi-cell StackBuf return needs a `let` binding")
+                };
+            }
+            self.lower_call(callee, args, dsts.len(), None, Some(dsts), false);
+        }
+    }
+
+    /// The runtime params, runtime args, and `Const`-substituted body of a call
+    /// to a user function: the ingredients for inlining. `None` for a builtin or
+    /// unknown callee, an arity mismatch, or an unresolved `Const` argument.
+    pub(super) fn specialized_body(&self, callee: &str, args: &[Expr]) -> Option<SpecializedBody> {
+        let def = self.defs.get(callee)?;
+        if args.len() != def.params.len() {
+            return None;
+        }
+        let mut body = def.body.clone();
+        let (mut rt_params, mut rt_args) = (Vec::new(), Vec::new());
+        for ((p, &is_const), a) in def.params.iter().zip(&def.const_params).zip(args) {
+            if !is_const {
+                rt_params.push(p.clone());
+                rt_args.push(a.clone());
+                continue;
+            }
+            let c = self.const_arg(a)?;
+            body = subst_stmts(&body, p, &c);
+        }
+        Some((rt_params, rt_args, body, def.n_ret))
+    }
+
+    /// If `callee` declares `Const` parameters, monomorphize: the constant
+    /// arguments (literals, `GEN ** k`, or literal-bound names) substitute into
+    /// a copy of the callee, queued once per distinct constant tuple and named
+    /// `callee__L5_G3`-style, and only the runtime arguments remain.
+    /// A callee's declared return shapes, looked up wherever it lives: an
+    /// ordinary definition sits in `defs`, while a `Const` specialization is
+    /// registered by [`Self::specialize`] in the queue under its mangled name and
+    /// never reaches `defs`. A dispatched `match_range` names specializations, so a
+    /// check that consults only `defs` silently passes on every one of them.
+    /// How many arguments `callee` takes, looked up the same way its return
+    /// shapes are, so a specialization or a generated helper answers too.
+    pub(super) fn arity_of(&self, callee: &str) -> Option<usize> {
+        self.defs
+            .get(callee)
+            .map(|d| d.params.len())
+            .or_else(|| self.queue.iter().find(|f| f.name == callee).map(|f| f.params.len()))
+    }
+
+    pub(super) fn return_shapes_of(&self, callee: &str) -> Option<Vec<ReturnShape>> {
+        self.defs.get(callee).map(|d| d.return_shapes.clone()).or_else(|| {
+            self.queue
+                .iter()
+                .find(|f| f.name == callee)
+                .map(|f| f.return_shapes.clone())
+        })
+    }
+
+    /// Consume the [`RetBind`] a single-value inlined tail return recorded,
+    /// for a call in EXPRESSION position (embedded in arithmetic, a store
+    /// RHS, a single-target match arm): there is no name to alias-bind, so an
+    /// aliased return materializes into a plain cell (free for a var / an
+    /// exp-0 g-address; one `MUL` for a shifted pointer). `dst` is the call's
+    /// destination cell, already written by a real call or a plain-scalar
+    /// return, so it is the fallback.
+    pub(super) fn take_inline_ret_cell(&mut self, dst: Off) -> Off {
+        match self.inline_stack_ret.take().and_then(|b| b.into_iter().next()) {
+            Some(RetBind::Gaddr(ga)) => self.materialize(ga),
+            Some(RetBind::Stack(base, size)) => {
+                if size != 1 {
+                    self.fail("a multi-cell StackBuf return needs a `let` binding, not an expression use")
+                };
+                // Through `word_src`, like every other read of a stack cell: the body may
+                // have filled this cell with a deferred copy or constant, which emits no
+                // instruction, and the raw cell would then be one no instruction writes.
+                // The `let` consumer follows the alias by taking a `Binding::Stack`
+                // ([`ret_binding`]), and an expression use has to agree with it.
+                self.word_src(base)
+            }
+            _ => dst,
+        }
+    }
+}
